프로젝트 파일 분리
