+++ FunchApp/Presentation/MultiProfileScene/MultiProfileListView.swift
@@ -7,68 +7,7 @@
 
 import SwiftUI
 
-//final class MultiProfileListViewBuilder {
-//    
-//    private var diContainer: DIContainer
-//    
-//    init(diContainer: DIContainer) {
-//        self.diContainer = diContainer
-//    }
-//    
-//    var body: some View {
-//        let viewModel = MultiProfileListViewModel(inject: <#T##DIContainer.Inject#>)
-//        let view = MultiProfileListView(viewModel: viewModel)
-//        
-//        return view
-//    }
-//}
 
-enum MultiProfileListPresentation: Hashable, Identifiable {
-    var id: Int { hashValue }
-    
-    case create
-    case home
-}
-
-
-final class MultiProfileListViewModel: ObservableObject {
-    
-    enum Action {
-        case load
-        case selection(Profile)
-        case presentation(MultiProfileListPresentation)
-    }
-    
-    @Published var presentation: MultiProfileListPresentation?
-    /// 프로필 목록
-    @Published var profiles: [Profile] = [.empty, .testableValue]
-    /// 유저가 선택한 프로필
-    @Published var selection: Profile?
-    
-    private var ineject: DIContainer.Inject
-
-    init(inject: DIContainer.Inject) {
-        self.ineject = inject
-    }
-    
-    func send(action: Action) {
-        switch action {
-        case .load:
-            self.profiles = ineject.userStorage.profiles.sorted { $0.createAt > $1.createAt }
-            
-            self.selection = ineject.userStorage.selectionProfile == nil
-            ? self.profiles.first
-            : self.ineject.userStorage.selectionProfile
-            
-        case let .selection(profile):
-            ineject.userStorage.selectionProfile = profile
-            self.selection = profile
-        
-        case let .presentation(presentation):
-            self.presentation = presentation
-        }
-    }
-}
 
 struct MultiProfileListView: View {
